+++ assets/javascripts/kitten/components/icons/arrow-icon.js
@@ -1,10 +1,10 @@
 import React from 'react'
 
-export const ArrowIcon = ({ className }) => {
+export const ArrowIcon = props => {
   return (
     <svg xmlns="http://www.w3.org/2000/svg"
          viewBox="0 0 6 8.99"
-         className={ className }>
+         { ...props }>
       <path d="M6 4.48L4.51 6 0 1.48 1.51 0z"/>
       <path d="M6 4.48L1.51 9 0 7.48 4.51 3z"/>
     </svg>