+++ CHANGELOG.md
@@ -6,21 +6,58 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: Fix wrong display for SearchInput component
 
+## [8.4.0] - 2017-03-08
+
+Features:
+- Accept more props on `ArrowIcon`.
+- Add new `LinkedinIcon` React component.
+- Add new `TwitterIcon` React component.
+- Add new `FacebookIcon` React component.
+- Add new `ButtonIcon` React component.
+- Add `defaultValue` to `Location`.
+
+Fixes:
+- `LocationInput` does not overwrite `onChange`.
+
+## [8.3.0] - 2017-03-08
+
+Features:
+- `SelectWithState` now accepts an `onInputChange` function.
+
+Fixes:
+- `SelectWithState` does not overwrite previous `inputProps`.
+
+## [8.2.1] - 2017-03-07
+
+Fixes:
+- Animation change when `Alert` closes.
+
+## [8.2.0] - 2017-03-06
+
+Features:
+- Add new `Alert` component.
+- Add new `CloseButton` component.
+- Add new `CrossIcon` component.
+
+Fixes:
+- `SelectWithState` defaults to `deleteRemoves` to `false`.
+- Use `CloseButton` component instead of simple button in `Popover`.
+
 ## [8.1.0] - 2017-03-06
 
-Feature:
+Features:
 - Add new `error` props on `RadioButton` and `Checkbox`.
 - Add new `LocationIcon` React component.
 - Add new `LocationInput` React component.
 
-Fix:
+Fixes:
 - Fix wrong padding for `InformationBox` component.
 - Use `handleChange` handler instead of `handleInput` in `TextInputWithLimit`
   to prevent React warning.
 
 ## [8.0.0] - 2017-02-28
 
-Breaking change:
+Breaking changes:
 - The `FormPhoneNumber` has new props.
   To be retro-compatible you should, in your app's components:
   - Rename `name` to `areaName`.
@@ -32,11 +69,11 @@ Breaking change:
 - Add `@include k-TextAreaWrapper;` in your SCSS if you use `TextInput` with
   `tag='texteara'`.
 
-Feature:
+Features:
 - Add gradient to textarea.
 - Add `limit` prop to textarea.
 
-Fix:
+Fixes:
 - Add blur event to `TextInput` component.
 - `TextInput` placeholder on Safari.
 - `TextInputWithUnit` does not pass over `digits` to its `TextInput`.
@@ -46,11 +83,11 @@ Fix:
 
 ## [7.1.0] - 2017-02-23
 
-Feature:
+Features:
 - Add digits prop to resize `TextInput`, `TextInputWithUnit` and
   `FormAmountAndCurrency`.
 
-Fix:
+Fixes:
 - Increase line-height on tiny `Label`.
 
 ## [7.0.0] - 2017-02-22